+++ src/nnet.c
@@ -222,7 +222,7 @@ int main_nnet(int argc, char* argv[argc])
 
 	if (train) {
 
-		long NO = config.get_no_odims(&config, NI, dims_in);
+		int NO = config.get_no_odims(&config, NI, dims_in);
 		long dims_out[NO];
 
 		complex float* out = load_cfl(filename_out, NO, dims_out);
